Dow not remove downloaded file if debug is enabled.

+++ src/main/java/de/cranix/services/EducationService.java
@@ -471,7 +471,9 @@ public class EducationService extends UserService {
                     }
                 }
         }
-        file.delete();
+	if (!logger.isDebugEnabled()) {
+            file.delete();
+	}
         return responses;
     }
